user create postman ok

+++ backend/database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('password');
             $table->string('sexe');
             $table->integer('age');
-            $table->decimal('size');
+            $table->integer('size');
             $table->decimal('weight');
             $table->string('objective');
             $table->string('activity');
